Remove admin hooks
